Add files via upload
Added first version of tic tac toe. I currently have no plans for this version, but I am trying to port it to JavaScript. This would be easy to accomplish, but I'm also trying to get it running on a GUI as well. First version with canvas had a good amount of issues. Looking for alternatives.